1. 대댓글 추가 1단계 2. RboardVO (int -> long 타입 수정)
끝.

+++ springedu/src/main/java/com/kh/portfolio/board/vo/RboardVO.java
@@ -13,9 +13,9 @@ import lombok.Data;
 public class RboardVO { 
 	
 	@NotNull
-	private int rnum; 					//	RNUM	NUMBER(10,0)	No		1	댓글 번호
+	private long rnum; 					//	RNUM	NUMBER(10,0)	No		1	댓글 번호
 	@NotNull
-	private int bnum;				  //	BNUM	NUMBER(10,0)	No		2	게시글 번호 / 최초 원글
+	private long bnum;				  //	BNUM	NUMBER(10,0)	No		2	게시글 번호 / 최초 원글
 	@NotNull
 	@Email
 //	@Pattern(regexp="\\w+@\\w+\\.\\w+(\\.\\w+)?",
